Production-ready infrastructure: errors, API, CLI, CI, safety, docs
- Unified error type (Rav2dError) replacing i32 errno and local Dav2dError
- Clean public API: internal modules hidden, selective re-exports
- Decoder pipeline wired: parse_obus integrated into gen_picture
- Working CLI: IVF demuxer, Y4M writer, decode loop with progress
- GitHub Actions CI: ubuntu/macOS matrix with fmt, clippy, test
- Safety audit: SAFETY comments on all Send/Sync impls, transmutes
  replaced with validated from_raw() helpers, pool leak fixed
- Criterion benchmarks for decoder operations
- Crate-level and per-type rustdoc documentation
- All TODO/FIXME markers resolved with AV2 spec references
- READMEs for all three crates

790 tests pass, clippy clean across workspace.

+++ crates/rav2d/src/env.rs
@@ -200,12 +200,12 @@ pub fn get_warpmv_2d(
 
     let mut res = MvXY {
         y: iclip(
-            apply_sign64(((yc.unsigned_abs() as i64 + rnd) >> shift) << not_epel, yc) as i32,
+            apply_sign64(((yc.unsigned_abs() as i64 + rnd) >> shift) << not_epel, yc),
             -max,
             max,
         ),
         x: iclip(
-            apply_sign64(((xc.unsigned_abs() as i64 + rnd) >> shift) << not_epel, xc) as i32,
+            apply_sign64(((xc.unsigned_abs() as i64 + rnd) >> shift) << not_epel, xc),
             -max,
             max,
         ),
